// src/api/backend/model/year.ts
/**
 * web-backend-swagger
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Year { 
    id: number;
    year: string;
    sealed: boolean;
    point_pad: number;
    active_orgs?: Array<number>;
    sum_points: number;
    tasks_cnt: number;
}